make make all make all
Add the builder components to the Makefile.

Leaves out jobsrv and worker for now.

Some of the Cargo.locks changed on me when building, but I'm assuming
this is ok.

Not all of the tests pass on my machine but I assume that's ok too.

Also update travis to install the protobuf crate and cache (some) apt
dependencies.