Adds `--bind` instead of watch, puts data in toml
This adds a `--bind` cli option for `hab start`. This will let you bind
a service group to a configuration key in the `bind` namespace.

`hab start --bind database:redis.default`

Would make `bind.has_database` and `bind.database` exist. `has_database`
is a boolean, that will be present when database data is available (ie:
the `redis.default` service group is present in the ring).

Additionally, this commit deletes the etcd and legacy user_config
module.

Signed-off-by: Adam Jacob <adam@chef.io>