Install core origin public key in devshell.
This is a temporary measure until public origin keys are downloaded on
package installations. Its implementation is non-straight forward
because we're caching the `/hab/cache/keys` directory with a mounted
data volume. When the Docker container comes up, this directory is bind
mounted over and the underlying directory contents is masked out.

Pull request: #457
Approved by: reset