Add a `plans/build_order.rb` which generates a build order for Plans.
The `plans/build_order.rb` is a Ruby program with no external library
dependencies, but it may call out to the `plans/build-base-plans.sh`
program which is assumed to be in the same directory.

The program is fed paths to `plan.sh` files on stdin which are used as
the input set of packages and dependencies. For example:

    find . -name plan.sh | ./plans/build_order.rb

Would find all plans in our project and determine the build order for
all of them. At the moment, only `core` origin Plans are considered (the
rest are ignored).

If the base set of packages have been built, as is usually the case, you
can add a `--without-base` flag to the program which, after sorting,
will remove all package identifers contained in the
`plans/build-base-plans.sh` program--which gets called with a
`$PRINT_IDENTS_ONLY` environment variable set. For example:

    find . -name plan.sh | ./plans/build_order.rb --without-base

Note that this program can be run on a Mac (which has a version of Ruby
already), in the devshell (assuming you install a Ruby package), or in a
Studio (assuming you build/install a `core/ruby`).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #464
Approved by: smith