[devshell] Prevent the `protobuf` crate binary from being mount-masked.
This change fixes an issue we had in the devshell when trying to build
all components. The `cargo install protobuf` command in the Dockerfile
honors the `$CARGO_HOME` environment variable override (as it should),
but that directory get overlay-mounted whenever a container is started
up which means that our lovingly compiled binaries get masked over.

Instead, we'll install this crate into the default location (under
root's home) and add that path to the environment's `$PATH`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #467
Approved by: adamhjk