Allow overriding of `pkg_origin` with HAB_ORIGIN
If you set the `HAB_ORIGIN` environment variable, we will use that when
building plans rather than the one specified in the package. This makes
it easy to create your own origin key, export `HAB_ORIGIN=foo`, and then
build any plan you want - with a resulting foo/PACKAGE as the result.

Signed-off-by: Adam Jacob <adam@chef.io>