[sup] Honor `HAB_RING` for `start` subcommand.
This change introduces an optional way to set the Ring name when a
Supervisor starts via setting a `HAB_RING` environment variable. The
environment variable will always be overrident if the user uses an
explicit `--ring` option on start, meaning that a CLI arugment is always
first priority and environment variable is secondary.

Current explicit behavior:

    hab-sup start --ring possums core/redis

New, alternative use with an environment variable:

    env HAB_RING=possums hab-sup start core/redis

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>