[sup] Honor `HAB_RING_KEY` containing key's content for `start` subcommand.
This change introduces a way to inject a Ring key to a Supervisor when
it starts via setting a new `HAB_RING_KEY` environment variable. Whereas
the previous `HAB_RING` environment variable contains the name of the
key (which is supposed to already exist on disk locally in the key
cache), the `HAB_RING_KEY` contains the contents of a key file itself.
This allows an operator to start a brand new Supervisor with the
following:

    env HAB_RING_KEY='SYM-SEC-1
    beyonce-20160504220722
    RCFaO84j41GmrzWddxMdsXpGdn3iuIy7Mw3xYrjPLsE=' hab-sup start core/redis

or alternatively:

    cat <<EOF > /tmp/key
    SYM-SEC-1
    beyonce-20160504220722

    RCFaO84j41GmrzWddxMdsXpGdn3iuIy7Mw3xYrjPLsE=
    EOF
    env HAB_RING_KEY="$(cat /tmp/key)" hab-sup start core/redis

or, even:

    env HAB_RING_KEY="$(curl https://extreme.trust/key)" hab-sup start core/redis

As before, there is a priority order in which CLI options and
environment variables are checked when setting the ring key on start:

1. The `--ring` option on the command line wins over any other setting
2. A set `$HAB_RING_KEY` environment variable is used next
3. A set `$HAB_RING` environment variable is used last
4. The Supervisor will start in an unencrypted mode

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>