[hab] Add `ring key import` & `ring key export` subcommands.
This change introduces two new subcommands which are intended to work
together to support a server-based workflow, where "server" may mean a
bare metal server, virtual machine, cloud instance, etc.

The first is a `hab ring key export` subcommand which outputs the latest
ring key's file contents to standard out. For example:

    hab ring key generate unicorns
    #=> Successfully generated ring key unicorns-20160505003452
    hab ring key export unicorns
    #=> SYM-SEC-1
    # unicorns-20160505003452
    #
    # 9vXIC0OQyfP2+MYakkhma9eU9oAM+xHPIkylFn4fXj4=

The flip side of this pairing is a `hab ring key import` subcommand
which consumes a standard input stream containing a ring key's file
contents, which it then writes to disk in the key cache with the correct
file name. Extending the example from above:

    cat <<EOF | hab ring key import
    SYM-SEC-1
    unicorns-20160505003452

    9vXIC0OQyfP2+MYakkhma9eU9oAM+xHPIkylFn4fXj4=
    EOF
    #=> Imported key unicorns-20160505003452

While this may not be terribly useful on a single system, pushing and
pulling ring keys across systems is now much easier with SSH pipes. For
example:

    # Push a ring key to another host
    hab ring key export unicorns | ssh node2 hab ring key import

    # Pull a ring key from another host
    ssh node2 hab ring key export unicorns | hab ring key import

    # Pull a ring key from another host and start a Supervisor
    env HAB_RING_KEY='$(ssh node1 hab ring key export unicorns)' hab-sup start core/redis

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>