Support `http_proxy` & `no_proxy` environment variables for Depot calls.
This feature adds support for using the `http_proxy` and
`no_proxy` environment variables when interacting with a Depot via the
`depot-client` component which powers the `hab` CLI and the `hab-sup`
Supervisor.

The raw environment variable parsing and interpretation is maintained in
the `core` component, which is then used by the `depot-client` when
creating `hyper::Client` instances. If the Depot URL's host/domain part
matches an entry in the `no_proxy` comma-separated list of domain
extensions, a proxy will not be used. Note that at the moment the
behavior when calling SSL-wrapped Depot instances is undefined and not
tested (which will happen later when this is officially supported).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>