[sup] Set appropriate `PATH` for `sh` & `bash` subcommands.
This change re-introduces the `bash` subcommand which was mistakenly
removed (sorry y'all) and also sets an appropriate `PATH` environment
variable.

The logic is similar to the path-setting logic of the `start` subcommand
except in this case, there is no Supervisor or service to have on path.
Instead, we determine a suitable Busybox path and append any `PATH`
entries at the end. This should make the Docker-wrapped services more
ergonomic when using `docker exec` or similar tricks.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #531
Approved by: smith