[hab] Add `hab origin key import` & `hab origin key export` subcommands.
This change introduces 2 new subcommands for origin keys:

* `hab origin key export`
* `hab origin key import`

The immediate use case to support Studio's upcoming key-injection
feature.

The `import` behavior is similar to the existing `hab ring key import`
subcommand, but the export takes a required option of `--type` which is
required to determine if a user intended to fetch the public or the
secret file.

With this new pairing of subcommands we can push/pull the secret origin
keys around, which are used to sign artifacts. For example:

```
hab origin key export --type secret unicorn \
  | ssh thevault hab origin key import
```

Also note that these subcommands also honor the `$FS_ROOT` environment
variable:

```
hab origin key export --type secret unicorn \
  | env FS_ROOT=/hab/studios/src hab origin key import
```

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>