Honor the `$HAB_ORIGIN` environment variable when launching a devshell.
This change allows a developer to export the `$HAB_ORIGIN` environment
variable on their workstation, which is propagated into a devshell when
`make shell` is used. The devshell's key cache is preserved/shared
between devshell via Docker mounted data volumes, meaning that you can
load a secret origin key in once and not re-copy it everytime you want
to launch a Studio.

For example, the following would work for a user that wants to build a
custom origin version of the `core/redis` Plan (this assumes that an
`fnichol` secret origin key is already present in `/hab/cache/keys`):

```
export HAB_ORIGIN=fnichol
make shell
hab-studio build plans/redis
```

The resulting artifact would be `./results/fnichol-redis-*.hart` signed
with the `fnichol` secret key.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>