[hab] Add `pkg exec` subcommand.
This change implements the `hab-bpm exec` in `hab` itself.

```
> hab pkg exec --help
hab-pkg-exec
Executes a command using the 'PATH' context of an installed package

USAGE:
    hab pkg exec [FLAGS] <PKG_IDENT> <CMD> [ARGS]

FLAGS:
    -h, --help       Prints help information
    -V, --version    Prints version information

ARGS:
    <PKG_IDENT>    A package identifier (ex: core/redis, core/busybox-static/1.42.2)
    <CMD>          The command to execute (ex: ls)
    <ARGS>...      Arguments to the command (ex: -l /tmp)

> hab pkg exec core/busybox-static ls -l /tmp
total 8
-rw-r--r--    1 root     root            75 May 19 04:17 core-20160423193745.pub
drwxr-xr-x    3 root     root          4096 May 19 04:17 npm-8-1a69e8d7

> hab pkg exec nope/nope nope
✗✗✗
✗✗✗ Cannot find a release of package: nope/nope
✗✗✗

> hab pkg exec core/vim viiiiiiim
✗✗✗
✗✗✗ `viiiiiiim' was not found on the filesystem or in PATH
✗✗✗

> hab pkg exec core/busybox-static/1.24.2/20160427212720 false
> echo $?
1
```

(Developer's note) this was more challenging a task than it first
appeared, partly because of some program argument pre-parsing to avoid
the cli parser trying to own all arguments, and partly because the
existing `hab-bpm exec` is precisely correct in a way that the
associated Rust code was not. For example the
`PackageInstall.runtime_path()` function was not as correct as the
implementation in `hab-bpm` which adds direct dependencies first, in
declared order, and then all remaining indirect dependencies in lexical
order. Also, the desired command doesn't necessarily have to live in the
target package, just somewhere in the resulting `PATH` set. Lastly, the
current `PATH` value is *not* added in order to preseve the strictly
correct `PATH` logic that a package might expect at runtime. Anyway,
cool times.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #545
Approved by: fnichol