[studio] Update Studio's install.sh to use `hab` only.
This change introduces a new `habitat-sh` S3 bucket with a copy of a
recent `core/hab-static` artifact. The installer has also been updated
to immediately upgrade to the current release of `core/hab-static` in an
effort to keep things current for as long as possible.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #556
Approved by: reset