Switch from Mustache to Handlebars
This commit moves the templating system from Mustache to Handlebars. It
brings us a couple of things:

1) Handlebars is a more flexible templating system; the Rust
implementation we are using will allow us to add custom handlers if we
want, making it easy to customize the templating system.

2) Handlebars uses the full Javascript true/false behavior on the
``{{if}}`` conditional. That means you can have values that use the
empty string evaluate correctly, making it much easier to write
templates.

3) Handlebars supports trimming whitespace with ``~``

4) Handlebars sets us up to support registering all the templates in a
plan, and letting you use some as partials.

5) The syntax is slightly more verbose than Mustache. For example:

```
{{#if cfg.your_value}}{{cfg.your_value}}{{/if}}
```

Versus

```
{{#cfg.your_value}}{{cfg.your_value}}{{/cfg.your_value}}
```

(Also, the Mustache version doesn't work unless the value of
``cfg.your_value`` is an array. This sucks. See reason 2)

Additionally, I replaced our use of the openssl hash function with the
same Blake2B hash we use elsewhere when deciding if we need to re-render
a configuration template.

Signed-off-by: Adam Jacob <adam@chef.io>