[hab] Add initial support for building a Mac `hab` binary.
This work is a result of a short research spike which looked into what a
`hab` binary for the Mac would look like. After finally having success
building the binary with the desired characteristics, I noted that our
build program--normally running inside a Studio--would be great at
capturing the build process. The change introduces a `mac/` subdirectory
under the hab component containing all necessary build details including
a script called `mac-build.sh` that wraps the main build program,
ensuring that all build dependencies are met before starting. Again, as
we don't have the benefit of a proper Mac build environment, we are
relying on Homebrew packages for the external linking libraries.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>