[core] Use a user-local cache directory if user is not root.
This change applies primarily to the `hab` binary, however any calling
code into the `core::fs::cache_*_path()` functions will get a path
either relative to the user running the process or the root of the
filesystem.

For example, if the root user (i.e. the effective ID of the process is
`0`), then the following default cache paths would be returned:

* `/hab/cache/artifacts`
* `/hab/cache/keys`
* `/hab/cache/src`

If the user is not root, in this example `jdoe` with a `$HOME` of
`/home/jdoe`, then the following default cache paths would be returned:

* `/home/jdoe/.hab/cache/artifacts`
* `/home/jdoe/.hab/cache/keys`
* `/home/jdoe/.hab/cache/src`

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #572
Approved by: adamhjk