Use `--listen-http` rather than `--listen-sidecar`
This PR updates the `--listen-sidecar` flag to be `--listen-http`. It
doesn't update all the internals to reflect it (since it's still the
"sidecar" module) - but I tried to change everywhere you reference the
configuration option.

It also is the first time the director has been updated post-handlebars
conversion, so it updated the Cargo.lock.

Signed-off-by: Adam Jacob <adam@chef.io>