Remove user-facing references to `hab-studio`, now under `hab studio`.
This change aims to tuck the Habitat Studio software under the
user-facing `hab` binary which owns the primary user experience.

The biggest change for Habitat developers is that `hab-studio` will no
longer be on `$PATH` in the devshell. To access it, just drop the dash,
and there you go: `hab-studio` becomes `hab studio`. See what we did
there?

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>