Move Makefile variable def's to the top.
Looks like there are parts of a Makefile that are parsed from top to
bottom. The last refactoring of the Makefile, which moved the
`IN_DOCKER` logic nearer to the bottom caused the `ifneq` blocks to not
be evaluated in time. The fix was to move this back up to the top which
sadly reduces the Makefile's readability, but at least it's correct
again. Sorry, y'all!

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #587
Approved by: juliandunn