[studio] Add a `hab` user & group to default Studio type.
This change coupled with #617 should allow a user to enter a default
Studio and immediately run:

    hab start core/redis

without issues.

Neither the Supervisor nor the `hab` CLI is responsible for adding users
and/groups (at least at the moment) meaning that if a service package
uses a non-privileged user to run under, this could result in a failure.
Further work can be done to have the package announce its user
requirements and have the Supervisor at least detect-and-fail is a user
is not present (minimal solution) or attempt to add a user/group (a
delightful solution).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>