[http-client] Drop support for http_proxy pending upstream support.
This change is admittedly a little sad, but is pragmatic in the short
term. The current releases of the Hyper crate (which is the underlying
HTTP client implementation) doesn't provide an API to construct a client
with a particular SSL context *and* http_proxy support. As a secured
Depot communication will trump most other features, we decided to
disable http_proxy support until the team can get a change upstream to
support this desired setup.

The previous commit (and therefore this one) will serve as the rough
implementation of `new_hyper_client` when we can restore proxy support.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #652
Approved by: adamhjk