Make origins work with new apis
Creating origins and listing user origins now works with the new APIs.
Key creating and invitations aren't there yet.

Had to make it so anybody can do GET /depot/origins/x and get the
response, since doing that is how we determine whether or not an origin
is available.

Also:

* Remove the builderApi client library (that doesn't use tokens) and
  finish putting everything in the BuilderApiClient library
* Make the client-side pattern for origins match what we're using on the
  server now
* Some UI around uploading keys, not totally finished
* Remove some API calls around "projects" that don't currently have
  anything that exists in the real API

Signed-off-by: Nathan L Smith <smith@chef.io>