Create `core/hab-bintray-publish` for Bintray publishing, right?
This change adds an additional piece of support software, under
`support/bintray-publish`, which automates the delivery of `hab` CLI
artifacts and Studio Docker images to Bintray for initial consumption.

As all plans, this can be build with:

    hab pkg build ./support/bintray-publish

Which produces a `core/hab-bintray-publish` package. There are 2
programs provided:

* `publish-hab` - used to take a Habitat package of `hab` and publish it
  to Bintray
* `publish-studio` - used to create and push a Docker image of a
  pre-created Studio instance

Both of these programs require login credentials to do their work and
these are consumed via environment variables. They will both fail if any
of the required environment variables are not set, but for clarity, the
following are used:

* `BINTRAY_USER` - username
* `BINTRAY_KEY` - an API key for the username
* `BINTRAY_PASSPHRASE` - the passphrase for the GPG-signing key on
  the Bintray platform

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>