[hab] Move `core/hab-static` to `core/hab`.
This change promotes the `core/hab` package to the statically built
binary. The glibc/dynamically linked version is moved to
`core/hab-dynamic` and may be removed in the future if it is not used.

Making this subtle change should reduce any confusion for users when
they attempt to download the `core/hab` package. Now they will get what
they (most likely) expected. Additionally, this now matches the Mac
version package identifier.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>