[hab] Update the `install.sh` to gpg-verify the published shasums.
This sets the `install.sh` script up to pull from Bintray when we cut
over. For the time being, this script is pulling from an S3 bucket which
contains the same files that are present on Bintray. Until we cut over,
pushing any new version of `core/hab` will again involve updating the
URL, and that will need to happen at least one more time.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #728
Approved by: smith