[core] Return Err in `PackageInstall::load` when pkg path not present.
Checking for the `/hab/pkgs` directory hasn't been a concern until we
started producing a portable `hab` CLI binary which doesn't necessarily
run inside a Habitat package itself. The package install detection logic
was initially running under the assumption that it was a Supervisor and
hence properly installed as a Habitat package. Then our glibc/dynamic
version of `hab` had the same set of assumptions (i.e. that it was
running in a Habitat package, hence the parent directories would exist).

This fix performs an early return if the package root path does not
exist which is accurate: no Habitat packages will exist so there's no
more detection work to do!

Closes #736

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #741
Approved by: fnichol