[plans] Add `unzip` to base plan set.
This change adds the `core/unzip` package to the dependencies for
`core/hab-plan-build` (i.e. the "build" program) with the result that
`unzip` will be present on `$PATH` for Plan authors.

Adding a new dependency to the build toolchain also meant updating the
base build set. Additionally, as `core/bash-static` and
`core/coreutils-static` are no longer in the build toolchain critical
path (from the retirement of `hab-bpm`), these were dropped from the
base build set.

Closes #591

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>