[hab] Fix `pkg hash` thread panic & re-format CLI parsing code.
This change fixes a mismatch argument value in the `hab pkg hash`
subcommand (`TARGET` vs. `SOURCE`).

Aditionally, the subcommand ordering the `cli.rs` and `main.rs` were
re-sorted in order to make the fake fuzzy matching code easier to manage
and to facilitate side-by-side viewing of the 2 source files.

Finally, the auth token parsing/lookup logic was refactored into a
function, similar to the origin parsing/lookup logic. This refactoring
adds an `auth_token` configuration value that can be set in a user's
`etc/cli.toml` configuration file. A later change may add an additional
question to the `hab cli setup` in order to populate this automatically.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>