[plan-build] Ensure that `HAB_BIN` even when `NO_INSTALL_DEPS` is set.
This is a change from former behavior when `hab` was used to only
install package dependencies. Now it is used in every run to hash all
the files and to sign that `FILES` metadata, and to sign the resulting
package artifact. Consequently, the `$NO_INSTALL_DEPS` environment
variable is used in the `_install_dependency()` function to decide
whether or not to install external packages.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>