Add support for Marathon (Mesos) applications with hab-pkg-mesosize
Mesos can use docker but the native format is a filesystem sandbox with cgroups and namespaces. hab-pkg-mesosize creates a mostly empty hab-studio with the new 'bare' format and packages it into a tarball. hab-pkg-mesosize provides the JSON syntax (with cpu, mem, # of instances and disk) for loading the application into Marathon via the UI. The tarball still needs an accessible URL, that may be done manually or investigated in a later PR for pushing it into Mesos' artifact store. Networking is not addressed in the current format either.