Updates to Travis for website deploy
* We've switched from pro to non-pro travis, so all of the secure keys
  have been regenerated
* Update the bucket policy to add the thing that was causing the deploy
  to the new bucket on the habitat org to fail
* Remove the slack notification, because it's disabled anyway (Homu
  posts notifications to Slack instead)
* Set `path_style = false` in the s3_sync; `true` is only needed if you
  have special characters in your bucket name; we don't
* Remove the Travis deploy for the bucket on the CIA account (currently
  the active website.)

Deploying to the new bucket will not work until the terraform change to
the bucket policy is applied.

This makes it so the website is not deployed to the bucket it's being
served from now, so the Fastly service will need to be pointed at the
new bucket once the deploy succeeds.

Signed-off-by: Nathan L Smith <smith@chef.io>