[hab] Make the `install.sh` script more portable with BusyBox userland.
This change aims to make the gzip'd tarball extraction more portable in
non-GNU userlands by first using `zcat` and then piping the stream to
`tar` for extraction. This change allows for a better out-of-the-box
experience in minimal distributions such as the BusyBox Docker image,
Alpine Linux, etc.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #982
Approved by: smith