[tf] Update bootstrap.sh to use Bintray assets.
This change updates the Terraform bootstrapping script to download the
initial `hab` from the Bintray platform rather than the previous S3
bucket. The S3 bucket currently in use is targetted for retirement and
should no longer be used if possible.

Due the format difference in the Bintray asset (that is, it is a
stripped binary-only archive), the tar extraction logic also had to be
updated.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>